\name{SIDES-package}
\alias{SIDES-package}
\alias{SIDES-package}
\docType{package}
\title{
Subgroup Identification Based on Differential Effect Search
}
\description{
Provides function to apply "Subgroup Identification based on Differential Effect Search" (SIDES) method proposed by Lipkovich et al. (2011) <doi:10.1002/sim.4289>.
}
\details{
\tabular{ll}{
Package: \tab SIDES\cr
Type: \tab Package\cr
Version: \tab 1.16\cr
Date: \tab 2021-05-16\cr
License: \tab GPL-3\cr
}
}
\author{
Marie-Karelle Riviere-Jourdan <eldamjh@gmail.com>
}
\references{
Ilya Lipkovich, Alex Dmitrienko, Jonathan Denne and Gregory Enas. Subgroup identification based on differential effect search - A recursive partitioning method for establishing response to treatment in patient subpopulations. Statistics in Medicine, 2011. <doi:10.1002/sim.4289>
}
